backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Vector Database
weaviate-client==3.25.3

# AI/ML
transformers==4.35.2
torch>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# HTTP
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Redis (Rate Limiting & Caching)
redis==5.0.1
fastapi-limiter==0.1.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Kaggle Integration
kaggle>=1.6.0
pandas>=2.3.0
